+++ pkg/streamingpromql/planning/core/vector_selector.go
@@ -135,7 +135,7 @@ func MaterializeVectorSelector(v *VectorSelector, _ *planning.Materializer, time
 		Subsets:                  subsets,
 	}
 
-	return planning.NewSingleUseOperatorFactory(selectors.NewInstantVectorSelector(selector, params.MemoryConsumptionTracker, params.QueryStats, v.ReturnSampleTimestamps, v.ReturnSampleTimestampsPreserveHistograms)), nil
+	return planning.NewSingleUseOperatorFactory(selectors.NewInstantVectorSelector(selector, params.MemoryConsumptionTracker, v.ReturnSampleTimestamps, v.ReturnSampleTimestampsPreserveHistograms)), nil
 }
 
 func (v *VectorSelector) ResultType() (parser.ValueType, error) {